+++ lib/validb/file_system_logger_worker.rb
@@ -1,0 +1,10 @@
+module Validb
+  class FileSystemLoggerWorker
+    include SidekiqStatus::Worker
+
+    def perform(message)
+      logger = Validb::Logger::FileSystem.new
+      logger.out(message)
+    end
+  end
+end