+++ lib/validb/model_validator.rb
@@ -1,15 +1,11 @@
 module Validb
   class ModelValidator
+    include SidekiqStatus::Worker
 
-    def initialize(params, logger)
-      @params = params
-      @batcher = Validb::Batcher.new(logger)
-    end
-
-    def validate(model)
-      $stdout.print "\nChecking #{model}(#{model.table_name}) (#{model.count} records)"
-      model.find_in_batches(batch_size: @params.batch_size) do |record_batch|
-        @batcher.validate(record_batch)
+    def perform(model_name, batch_size)
+      model_name.constantize.select(:id).find_in_batches(batch_size: batch_size) do |record_batch|
+        model_ids = record_batch.map(&:id)
+        Validb::Batcher.perform_async(model_name, model_ids)
       end
     end
   end