+++ spec/spec_helper.rb
@@ -1,10 +1,17 @@
 require File.expand_path(File.dirname(__FILE__) + '/../lib/validb')
 
 require "pry"
+require "sidekiq"
+require 'sidekiq/testing'
+require 'sidekiq_status'
 
 RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
+
+  config.after(:all) do
+    SidekiqStatus::Container.delete
+  end
 end
 
 RSpec::Matchers::OperatorMatcher.register(ActiveRecord::Relation, '=~', RSpec::Matchers::BuiltIn::MatchArray)