+++ spec/validb/version_spec.rb
@@ -2,6 +2,6 @@ require 'spec_helper'
 
 describe Validb do
   it "returns the version" do
-    Validb::VERSION.should == "1.2.1"
+    Validb::VERSION.should == "2.0.0"
   end
 end